sparate product form list
