+++ src/components/Product/Form/types.ts
@@ -1,0 +1,6 @@
+export type FormInputType = {
+  courseName: string;
+  categoryId: number;
+  name: string;
+  price: number;
+};
